Backend/requirements.txt
apturl==0.5.2
arrow==1.2.1
asgiref==3.3.0
async-generator==1.10
attrs==21.4.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blessed==1.19.0
blinker==1.4
boto3==1.20.26
botocore==1.23.26
Brlapi==0.7.0
bs4==0.0.1
cache==1.0.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2020.11.8
cffi==1.15.0
channels==3.0.5
chardet==3.0.4
charset-normalizer==2.0.9
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.3
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.2
django-filter==21.1
django-mysql==3.9.0
django-nine==0.2.5
django-picklefield==3.0.1
django-q==1.3.9
django-redis==5.1.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
duplicity==0.8.12.0
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
firebase-admin==6.1.0
Flask==2.2.2
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.1
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
h11==0.12.0
httplib2==0.22.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==5.0.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
msgpack==1.0.5
mysql==0.0.3
mysqlclient==2.0.1
Naked==0.1.32
netifaces==0.10.4
numpy==1.23.5
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.5.2
paramiko==2.6.0
pdfkit==1.0.0
pexpect==4.6.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
pymongo==3.12.1
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-fcm==0.1
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requests-unixsocket==0.2.0
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
SecretStorage==2.3.1
selenium==4.1.0
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.43
sqlparse==0.2.4
status==0.2.5
systemd-python==234
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.2
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.2.2
wsproto==1.0.0
xkit==0.0.0
XlsxWriter==3.0.3
zipp==3.10.0
zope.interface==6.0
